requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
openpyxl
gtts
pillow
streamlit
plotly
google-generativeai
python-multipart